+++ src/Services/Converters/ExceptionConverterInterface.php
@@ -14,9 +14,16 @@ interface ExceptionConverterInterface
 {
 
     /**
+     * Convert the exception to an array of message parameters:
+     *
+     *  * code - the HTTP status code
+     *  * data.message - the error message to display
+     *  * data.errors - an array of key -> value pairs of error data e.g. fields
+     *  * data.XXX - any other elements to be added to the response
+     *
      * @param Exception $e
      *
-     * @return array An array containing "data.error" - the error message and "code" the HTTP status code
+     * @return array
      */
     public function convert(Exception $e): array;
 }